refactor: resolve follow-up package split gaps
