+++ libraries/typescript/packages/cli/package.json
@@ -55,7 +55,8 @@
     "commander": "^14.0.2",
     "dotenv": "^17.3.1",
     "esbuild": "0.28.1",
-    "mcp-use": "workspace:*",
+    "@mcp-use/client": "workspace:*",
+    "@mcp-use/server": "workspace:*",
     "open": "^11.0.0",
     "tar": "^7.5.16",
     "tsx": "^4.21.0",
