+++ libraries/typescript/packages/inspector/package.json
@@ -80,7 +80,9 @@
   },
   "peerDependencies": {
     "express": "^4.21.2 || ^5.0.0",
-    "mcp-use": ">=1.32.1-canary.0",
+    "@mcp-use/client": ">=1.0.0",
+    "@mcp-use/server": ">=1.0.0",
+    "@mcp-use/react": ">=1.0.0",
     "react": "^18.0.0 || ^19.0.0",
     "react-dom": "^18.0.0 || ^19.0.0",
     "react-router": "^7.12.0"
@@ -114,7 +116,9 @@
     "hono": "^4.12.12",
     "lucide-react": "^0.562.0",
     "markdown-to-jsx": "^9.7.4",
-    "mcp-use": "workspace:*",
+    "@mcp-use/client": "workspace:*",
+    "@mcp-use/server": "workspace:*",
+    "@mcp-use/react": "workspace:*",
     "motion": "^12.34.2",
     "next-themes": "^0.4.6",
     "open": "^11.0.0",
