+++ libraries/typescript/packages/inspector/src/client/App.tsx
@@ -9,7 +9,7 @@ import {
   LocalStorageProvider,
   McpClientProvider,
   type McpServer,
-} from "mcp-use/react";
+} from "@mcp-use/react";
 import { useEffect, useMemo, useRef } from "react";
 import { Route, BrowserRouter as Router, Routes } from "react-router";
 import { toast } from "sonner";
