+++ libraries/typescript/packages/inspector/src/client/components/LayoutContent.tsx
@@ -1,4 +1,4 @@
-import type { McpServer } from "mcp-use/react";
+import type { McpServer } from "@mcp-use/react";
 import type { ReactNode, RefObject } from "react";
 import { useInspector } from "@/client/context/InspectorContext";
 import type { TabType } from "@/client/context/InspectorContext";
